Define the Action type.
An Action is a composable IO action that takes a preset as an input
and returns a (possibly) modified preset. Most command line switches
should correspond to an Action. Thus, the list of switches passed to
the program corresponds to a sequence of Actions that will be applied
to a Preset and whose output can be written to file.