Fix accidental double-decoding of some data.
For presets with version 2.2, any embedded pattern is saved in the
"Texture/Pattern/Pattern" parameter as a binary value. In this
case, there will also be a matching "Texture/Pattern/PatternMD5"
parameter which contains an MD5 checksum. However, both these
values will be base64-encoded twice.

To get the actual data, we must decode these values a second time.
Similarly when we saving the preset, we have to encode the data
twice before inserting it into the settings document.

Prior to this change, all binary data was decoded twice (if possible).
However, we only want to decode twice in the specific situation above.